show the same session list in multiple places
