+++ webapp/src/components/LandingPage.vue
@@ -13,35 +13,11 @@
 			.header
 				h3 Featured Sessions
 				router-link(:to="{name: 'schedule'}") full schedule
-			.sessions
-				.session(v-for="{session, state}, index of featuredSessions", :class="{live: state.isLive}")
-					.speaker-avatars
-						template(v-for="speaker of session.speakers")
-							img(v-if="speaker.avatar", :src="speaker.avatar")
-							identicon(v-else, :id="speaker.name")
-					.info
-						.title-time
-							.title {{ $localize(session.title) }}
-							.time {{ state.timeString }}
-						.speakers-room
-							.speakers {{ session.speakers ? session.speakers.map(s => s.name).join(', ') : '' }}
-							.room {{ $localize(session.room.name) }}
+			session-list(:sessions="featuredSessions")
 			.header
 				h3 Sessions Happening Soon
 				router-link(:to="{name: 'schedule'}") full schedule
-			.sessions
-				.session(v-for="{session, state}, index of nextSessions", :class="{live: state.isLive}")
-					.speaker-avatars
-						template(v-for="speaker of session.speakers")
-							img(v-if="speaker.avatar", :src="speaker.avatar")
-							identicon(v-else, :id="speaker.name")
-					.info
-						.title-time
-							.title {{ $localize(session.title) }}
-							.time {{ state.timeString }}
-						.speakers-room
-							.speakers {{ session.speakers ? session.speakers.map(s => s.name).join(', ') : '' }}
-							.room {{ $localize(session.room.name) }}
+			session-list(:sessions="nextSessions")
 		.speakers
 			.header
 				h3 Our Speakers
@@ -61,9 +37,10 @@ import api from 'lib/api'
 import moment from 'lib/timetravelMoment'
 import Identicon from 'components/Identicon'
 import MarkdownContent from 'components/MarkdownContent'
+import SessionList from 'components/SessionList'
 
 export default {
-	components: { Identicon, MarkdownContent },
+	components: { Identicon, MarkdownContent, SessionList },
 	props: {
 		module: Object
 	},
@@ -81,17 +58,16 @@ export default {
 			if (!this.sessions) return
 			// return this.sessions.filter(session => session.featured)
 			// TODO remove mock data
-			return this.sessions.slice(0, 3).map(session => ({session, state: this.getSessionState(session)}))
+			return this.sessions.slice(0, 3)
 		},
 		nextSessions () {
 			if (!this.sessions) return
 			// current or next sessions per room
 			const sessions = []
-			// TODO filter out sessions with no venueless room?
 			for (const session of this.sessions) {
 				if (!session.room) continue
-				if (session.end.isBefore(this.now) || sessions.reduce((acc, s) => s.session.room === session.room ? ++acc : acc, 0) >= 2) continue
-				sessions.push({session, state: this.getSessionState(session)})
+				if (session.end.isBefore(this.now) || sessions.reduce((acc, s) => s.room === session.room ? ++acc : acc, 0) >= 2) continue
+				sessions.push(session)
 			}
 			return sessions
 		},
@@ -112,24 +88,6 @@ export default {
 			focus: 'center',
 			// padding: '16px 0'
 		}).mount()
-	},
-	methods: {
-		getSessionState (session) {
-			if (session.start.isBefore(this.now)) {
-				return {
-					isLive: true,
-					timeString: 'live'
-				}
-			}
-			// if (session.start.isBefore(this.now)) {
-			// 	return {
-			// 		timeString: 'starting soon'
-			// 	}
-			// }
-			return {
-				timeString: moment.duration(session.start.diff(this.now)).humanize(true).replace('minutes', 'mins')
-			}
-		}
 	}
 }
 </script>
@@ -165,69 +123,6 @@ export default {
 		h3
 			margin: 0
 			line-height: 56px
-	.sessions
-		display: flex
-		flex-direction: column
-		border: border-separator()
-		.session
-			height: 56px
-			position: relative
-			padding: 4px 0
-			box-sizing: border-box
-			display: flex
-			cursor: pointer
-			&:not(:last-child)
-				border-bottom: border-separator()
-			&:hover
-				background-color: $clr-grey-100
-			.speaker-avatars
-				flex: none
-				> *:not(:first-child)
-					margin-left: -28px
-				img
-					background-color: $clr-white
-					border-radius: 50%
-					height: 48px
-					width: 48px
-					margin: 0 8px 0 4px
-					object-fit: cover
-			.info
-				flex: auto
-				min-width: 0
-				display: flex
-				flex-direction: column
-				justify-content: space-between
-			.title-time
-				flex: none
-				display: flex
-				align-items: center
-				overflow: hidden
-			.title
-				flex: auto
-				// font-weight: 600
-				// line-height: 32px
-				line-height: 28px
-				ellipsis()
-			.time
-				flex: none
-				background-color: $clr-blue-grey-500
-				color: $clr-primary-text-dark
-				padding: 2px 4px
-				margin: 0 4px 4px
-				border-radius: 4px
-			.speakers-room
-				display: flex
-				align-items: baseline
-				margin-right: 4px
-			.speakers, .room
-				flex: 1
-				color: $clr-secondary-text-light
-				ellipsis()
-			.room
-				margin-left: 4px
-				text-align: right
-			&.live .time
-				background-color: $clr-danger
 	.speakers-list
 		display: flex
 		flex-wrap: wrap
